Split classify_corpus.rs to clear max-lines (parts moved to a subdirectory so Cargo's example autodiscovery does not treat them as separate binary targets)
