+++ crates/almide-mir/examples/classify_corpus_parts/classify_corpus_b.rs
@@ -1,0 +1,498 @@
+
+fn discover_self_modules(
+    path: &Path,
+    prog: &almide_lang::ast::Program,
+) -> Vec<(String, almide_lang::ast::Program, bool)> {
+    // Resolve a `self.<submodule>` OR an EXTERNAL (non-stdlib) package import — so a
+    // cross-module file works under `almide run`-equivalent resolution (incl. fetched deps),
+    // not just self-imports. A lone / stdlib-only file stays a strict no-op.
+    let needs_resolve = prog.imports.iter().any(|d| {
+        matches!(d, almide_lang::ast::Decl::Import { path, .. }
+            if path.first().map(|s| {
+                let s = s.as_str();
+                s == "self" || !almide_lang::stdlib_info::is_stdlib_module(s)
+            }).unwrap_or(false))
+    });
+    if !needs_resolve {
+        return Vec::new();
+    }
+    let deps = dep_paths_for(path);
+    match almide::resolve::resolve_imports_with_deps(&path.to_string_lossy(), prog, &deps) {
+        Ok(resolved) => resolved
+            .modules
+            .into_iter()
+            .map(|(name, p, _pkg, is_self)| (name, p, is_self))
+            .collect(),
+        Err(_) => Vec::new(),
+    }
+}
+
+/// The mangled flat name a user-module function gets when resolved to a user `CallFn`
+/// (`bindgen` + `get_str` → `almide_rt_bindgen_get_str`) — the v1 analogue of v0's
+/// `ir_link_flatten` module-fn renaming.
+fn user_module_fn_name(module: &str, func: &str) -> String {
+    format!("almide_rt_{}_{}", module.replace('.', "_"), func.replace('.', "_"))
+}
+
+/// Build the set of NATIVE-FFI function keys over the LINKED IR: functions that TRANSITIVELY
+/// reach a STRUCTURAL native root — a root being EITHER an `@extern(rust/rs)` declaration (no
+/// wasm form: porta `wasm_rt.almd` `wt_*`, almide-sqlite `native_*`) OR a DIRECT call to a
+/// permanently-no-wasm stdlib effect (`net.*`, `process.exec/exit/run`, `http.request`). Such a
+/// function can NEVER lower to wasm; its wall is STRUCTURAL, not a v1 lowering gap, so it must be
+/// EXCLUDED from the wall=0 metric exactly like the already-excluded `@extern(wasm)` WASI imports
+/// (those lower via `extern_wasm_target` and never wall — they are not roots here).
+///
+/// SOUNDNESS (do-NOT-over-exclude): ONLY those two root shapes seed the set. A function that walls
+/// on a PURE lowering gap (heap-result return, etc.) and never transitively reaches a root stays
+/// REAL — its gap is NOT hidden. `process.args` is deliberately EXCLUDED from roots (WASI
+/// `args_get` exists ⇒ uncertain ⇒ conservative REAL). `CallTarget::Method`/`Computed` are
+/// unresolved ⇒ no edge, no root (conservative REAL). Bare `FnRef`/`ClosureCreate` references are
+/// NOT root contributions.
+///
+/// Keys match the edge names `resolve_user_module_calls` produced: file functions (`ir.functions`)
+/// by plain name; module functions (`ir.modules[].functions`) by `user_module_fn_name(module, func)`
+/// — exactly the `CallTarget::Named` a resolved user-module call carries, so edges resolve.
+fn compute_native_ffi_set(ir: &almide_ir::IrProgram) -> HashSet<String> {
+    use almide_ir::CallTarget;
+    // 1) NODE KEYS over BOTH function sets, paired with the body to scan for edges/roots.
+    let mut nodes: Vec<(String, &almide_ir::IrFunction)> = Vec::new();
+    for f in &ir.functions {
+        nodes.push((f.name.as_str().to_string(), f));
+    }
+    for m in &ir.modules {
+        for f in &m.functions {
+            nodes.push((user_module_fn_name(m.name.as_str(), f.name.as_str()), f));
+        }
+    }
+    // Root-(a): an `@extern` decl whose target is `rust`/`rs`/`c` (NOT `wasm`/`ts` — those lower
+    // to a WASI/browser import via `extern_wasm_target` and never wall; this exclusion is the
+    // precedent). `c` is a C-library link (`@extern(c, "m", "sqrt")` — extern_c_test, whose header
+    // says "wasm:skip — @extern(c) not available in WASM"): structurally native, same as rust/rs.
+    let is_native_extern = |f: &almide_ir::IrFunction| {
+        f.extern_attrs.iter().any(|a| matches!(a.target.as_str(), "rust" | "rs" | "c"))
+    };
+    // Per-body collector: forward `Named` call edges + whether the body DIRECTLY calls an
+    // enumerated permanently-no-wasm stdlib effect (root-(b)). A remaining `CallTarget::Module`
+    // is ALWAYS a stdlib call here (user modules were already rewritten to `Named` by
+    // `resolve_user_module_calls`), so this matches only the genuine stdlib effect roots.
+    struct Collector {
+        edges: Vec<String>,
+        native_call: bool,
+    }
+    impl almide_ir::visit::IrVisitor for Collector {
+        fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
+            use almide_ir::IrExprKind::{Call, TailCall};
+            let target = match &e.kind {
+                Call { target, .. } | TailCall { target, .. } => Some(target),
+                _ => None,
+            };
+            if let Some(t) = target {
+                match t {
+                    CallTarget::Named { name } => self.edges.push(name.as_str().to_string()),
+                    CallTarget::Module { module, func, .. } => {
+                        let (m, fname) = (module.as_str(), func.as_str());
+                        // net.* (any func) / the no-wasm process fns / http.request / zlib.* —
+                        // the tight, enumerated no-wasm set. process.args is EXCLUDED (WASI
+                        // args_get exists AND v0's emit_wasm implements it — calls_process.rs
+                        // handles exactly exit/stdin_lines/args). spawn/kill/is_alive/
+                        // exec_status/env have NO v0 wasm form (the fixture headers declare
+                        // them native-only: "wasm:skip — process.env/spawn/kill are
+                        // native-only" / "process.exec_status is native-only"), and WASI
+                        // preview1 has no child-process API — structural, not a v1 gap.
+                        // zlib has NO v0 wasm runtime at all ("wasm:skip — OS/native-only").
+                        // http.serve is a TCP LISTENER ("wasm:skip — http.serve is native-only",
+                        // effect_intrinsic_tail_test) — the same no-wasm class as net.*.
+                        // random is deliberately NOT here: v0's emit_wasm implements it over
+                        // WASI random_get (calls_random.rs) — a REAL v1 gap.
+                        // testing.assert_throws needs `std::panic::catch_unwind` (runtime/rs/src/
+                        // testing.rs) — WASM's `unreachable` trap is NOT catchable (no unwind
+                        // mechanism in the WASI MVP ABI v0 targets). v0's OWN emit_wasm has no
+                        // wasm form either (calls_p2.rs's `assert_throws` arm is native-only) —
+                        // the fixture header says so verbatim ("wasm:skip — WASM cannot catch
+                        // panics"), matching CHANGELOG.md and wasm_dispatch_coverage_test.rs's
+                        // independent documentation of the same limitation. Structural, not a v1
+                        // lowering gap — the E3/E4/E5 native-root precedent.
+                        if m == "net"
+                            || m == "zlib"
+                            || (m == "process"
+                                && matches!(
+                                    fname,
+                                    "exec" | "exit" | "run" | "spawn" | "kill" | "is_alive"
+                                        | "exec_status" | "env"
+                                ))
+                            || (m == "http" && matches!(fname, "request" | "serve"))
+                            || (m == "testing" && fname == "assert_throws")
+                        {
+                            self.native_call = true;
+                        }
+                    }
+                    // Method / Computed: unresolved → no edge, no root (conservative REAL).
+                    _ => {}
+                }
+            }
+            almide_ir::visit::walk_expr(self, e);
+        }
+    }
+    // 2/3) Seed roots + record forward edges.
+    let mut native: HashSet<String> = HashSet::new();
+    let mut fwd: BTreeMap<String, Vec<String>> = BTreeMap::new();
+    for (key, f) in &nodes {
+        let mut c = Collector { edges: Vec::new(), native_call: false };
+        almide_ir::visit::IrVisitor::visit_expr(&mut c, &f.body);
+        if is_native_extern(f) || c.native_call {
+            native.insert(key.clone());
+        }
+        fwd.insert(key.clone(), c.edges);
+    }
+    // 4) PROPAGATE callee→caller to a fixpoint: a caller of a native fn is itself native.
+    loop {
+        let mut changed = false;
+        for (caller, callees) in &fwd {
+            if native.contains(caller) {
+                continue;
+            }
+            if callees.iter().any(|c| native.contains(c)) {
+                native.insert(caller.clone());
+                changed = true;
+            }
+        }
+        if !changed {
+            break;
+        }
+    }
+    native
+}
+
+/// Resolve a USER-package/-module call (`bindgen.get_str` via `import self as bindgen`,
+/// `self.classifier.classify`) to a real user `CallFn` (`CallTarget::Module` → `Named`,
+/// `almide_rt_<m>_<f>`). The MIR lowering then treats it as an ordinary user call instead of
+/// walling it as an opaque "impure stdlib Module" call. SOUNDNESS (caps): the resolved name
+/// carries NO dot, so the transitive caps gate analyzes it as a user call (via the in-profile
+/// map, or TAINTS the caller if the callee is a cross-file/unanalyzable definition) — NOT as a
+/// pure dotted stdlib call (`is_known_free`). A self-pkg call to an EFFECTFUL user fn thus
+/// surfaces its capability transitively, exactly like any user call — never the
+/// accept-but-unsafe omission the Module-call purity wall guarded against. STDLIB modules are
+/// NOT rewritten. No-op when there are no linked user modules.
+fn resolve_user_module_calls(ir: &mut almide_ir::IrProgram) {
+    use almide_ir::{CallTarget, IrExprKind, IrMutVisitor, walk_expr_mut};
+    use almide_lang::intern::sym;
+    let user_mods: BTreeMap<String, HashSet<String>> = ir
+        .modules
+        .iter()
+        .filter(|m| !almide_lang::stdlib_info::is_any_stdlib(m.name.as_str()))
+        .map(|m| {
+            (
+                m.name.as_str().to_string(),
+                m.functions.iter().map(|f| f.name.as_str().to_string()).collect(),
+            )
+        })
+        .collect();
+    if user_mods.is_empty() {
+        return;
+    }
+    struct Rw<'a> {
+        user_mods: &'a BTreeMap<String, HashSet<String>>,
+    }
+    impl IrMutVisitor for Rw<'_> {
+        fn visit_expr_mut(&mut self, e: &mut almide_ir::IrExpr) {
+            walk_expr_mut(self, e);
+            if let IrExprKind::Call { target, .. } = &mut e.kind {
+                if let CallTarget::Module { module, func, .. } = target {
+                    let (m, f) = (module.as_str(), func.as_str());
+                    if self.user_mods.get(m).is_some_and(|fs| fs.contains(f)) {
+                        *target = CallTarget::Named { name: sym(&user_module_fn_name(m, f)) };
+                    }
+                }
+            }
+        }
+    }
+    let mut rw = Rw { user_mods: &user_mods };
+    for func in &mut ir.functions {
+        rw.visit_expr_mut(&mut func.body);
+    }
+    for tl in &mut ir.top_lets {
+        rw.visit_expr_mut(&mut tl.value);
+    }
+    for m in &mut ir.modules {
+        for func in &mut m.functions {
+            rw.visit_expr_mut(&mut func.body);
+        }
+        for tl in &mut m.top_lets {
+            rw.visit_expr_mut(&mut tl.value);
+        }
+    }
+}
+
+/// Drive source → linked IR with NO `die()` — every failure becomes a value, so
+/// the sweep never aborts on a single bad file. Mirrors `emit_cert_from_source`'s
+/// pipeline (the same public frontend functions almide-interp uses). `path` is the
+/// file's location so `import self.<submodule>` resolves its sibling `src/*.almd`
+/// (the canonical driver discovery) — exactly the cut point `render_program` uses, so
+/// the wall report counts a cross-module function's REAL lowerability, not the
+/// missing-sibling artifact. A lone single file (no `import self.*`) is unchanged.
+fn source_to_ir(path: &Path, source: &str) -> FrontendOutcome {
+    let result = catch_unwind(AssertUnwindSafe(|| -> Result<almide_ir::IrProgram, String> {
+        let tokens = Lexer::tokenize(source);
+        let mut parser = Parser::new(tokens);
+        let mut prog = parser.parse().map_err(|e| format!("parse error: {e:?}"))?;
+        if !parser.errors.is_empty() {
+            return Err(format!("parse errors: {:?}", parser.errors));
+        }
+        let modules = discover_self_modules(path, &prog);
+        let canon = canonicalize::canonicalize_program(
+            &prog,
+            modules.iter().map(|(n, p, s)| (n.as_str(), p, *s)),
+        );
+        let mut checker = Checker::from_env(canon.env);
+        let diags = checker.infer_program(&mut prog);
+        let errors: Vec<_> = diags
+            .iter()
+            .filter(|d| d.level == almide_frontend::diagnostic::Level::Error)
+            .map(|d| d.message.clone())
+            .collect();
+        if !errors.is_empty() {
+            return Err(format!("type errors ({} diag)", errors.len()));
+        }
+        let mut ir = lower_program(&prog, &checker.env, &checker.type_map);
+        // Lower each resolved sibling MODULE into `ir.modules` — the SAME sequence the real
+        // driver runs (infer_module → per-module import table → lower_module → push), so a
+        // cross-module record/variant type reaches `build_record_layouts`. Non-bundled stdlib
+        // modules are skipped (their defs come from the runtime registry, not user lowering).
+        for (name, mod_prog, _is_self) in &modules {
+            if almide_lang::stdlib_info::is_stdlib_module(name)
+                && !almide_lang::stdlib_info::is_bundled_module(name)
+            {
+                continue;
+            }
+            let mut mod_prog = mod_prog.clone();
+            checker.infer_module(&mut mod_prog, name);
+            let self_name = checker.env.self_module_name.map(|s| s.to_string());
+            let import_table_name = self_name.as_deref().unwrap_or(name.as_str());
+            let (mod_table, _) = almide_frontend::import_table::build_import_table(
+                &mod_prog,
+                Some(import_table_name),
+                &checker.env.user_modules,
+            );
+            let saved_table = std::mem::replace(&mut checker.env.import_table, mod_table);
+            let mod_ir = almide_frontend::lower::lower_module(
+                name,
+                &mod_prog,
+                &checker.env,
+                &checker.type_map,
+                None,
+            );
+            checker.env.import_table = saved_table;
+            ir.modules.push(mod_ir);
+        }
+        // Resolve self-pkg / imported user-module calls to real user CallFns (Module → Named,
+        // `almide_rt_<m>_<f>`), so the MIR lowering treats them as ordinary user calls instead of
+        // walling them as opaque "impure stdlib Module" calls. SOUNDNESS: the resolved name has no
+        // dot, so the transitive caps gate analyzes it as a user call (in-profile map / taint),
+        // NOT a pure dotted stdlib call — a self-pkg call to an effectful user fn surfaces its
+        // capability transitively. No-op when there are no linked user modules.
+        resolve_user_module_calls(&mut ir);
+        optimize::optimize_program(&mut ir);
+        mono::monomorphize(&mut ir);
+        ir_link::ir_link(&mut ir);
+        // Transparent-newtype erasure LAST (post-link, pre-lowering) — the SAME pass the
+        // pipeline runs, so the caps mir == ir count sees the erased tree on both sides.
+        almide_mir::lower::erase_transparent_newtypes(&mut ir);
+        almide_mir::lower::fill_record_defaults(&mut ir);
+        almide_mir::lower::inline_pure_call_globals(&mut ir);
+        // #806 step 2 — the SAME small-scalar-fn inline the pipeline runs.
+        almide_mir::lower::inline_small_scalar_fns(&mut ir);
+        // C-132 move-mode write-back — the SAME pre-lowering rewrite the pipeline
+        // runs (see source_to_ir_with), so mir == ir on both sides.
+        almide_ir::mut_param::lower_mut_params_move_mode(&mut ir);
+        // Guard → if restructure — the SAME pre-lowering pass the pipeline runs.
+        almide_mir::lower::desugar_fn_body_guards(&mut ir);
+        almide_mir::lower::normalize_tail_err_raise_ifs(&mut ir);
+        almide_mir::lower::hoist_block_call_args(&mut ir);
+        almide_mir::lower::desugar_loop_early_returns(&mut ir);
+        almide_mir::lower::hoist_spread_call_bases(&mut ir);
+        almide_mir::lower::hoist_record_literal_args(&mut ir);
+        Ok(ir)
+    }));
+    match result {
+        Ok(Ok(ir)) => FrontendOutcome::Ir(ir),
+        Ok(Err(_reason)) => FrontendOutcome::Rejected,
+        Err(_) => FrontendOutcome::Panicked,
+    }
+}
+
+/// Count the `StringInterp` SITES in a body, split into `(proven, walled)`:
+///   - PROVEN (a): the interp DESUGARS and EVERY synthetic `<module>.to_string` it
+///     introduces is LINKABLE (in the self-host registry — `int`/`bool`). It folds to a
+///     fully-resolvable `__str_concat` / `int.to_string` / `bool.to_string` chain that
+///     renders to valid wasm and byte-matches v0.
+///   - WALLED (b): the interp is NON-desugarable (a part with no admitted `to_string`
+///     module ⇒ stays Opaque), OR it desugars but a synthetic `to_string` is UNLINKED
+///     (Float/compound — `float.to_string`/`list.to_string` are not self-hosted), so the
+///     enclosing function emits an unlinked call and the render wall rejects it. Either
+///     way: no invalid wasm. (The unlinked-call OCCURRENCES are also folded into (b) by
+///     the per-CallFn loop at the lowering site; this only counts the interp SITE once, as
+///     walled, so a Float interp does not get mis-bucketed as proven.)
+fn count_interp_sites(
+    body: &almide_ir::IrExpr,
+    linkable: &HashSet<String>,
+    registry: &almide_mir::lower::RecordLayouts,
+) -> (usize, usize) {
+    struct InterpCounter<'a> {
+        proven: usize,
+        walled: usize,
+        linkable: &'a HashSet<String>,
+        registry: &'a almide_mir::lower::RecordLayouts,
+    }
+    impl almide_ir::visit::IrVisitor for InterpCounter<'_> {
+        fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
+            if let almide_ir::IrExprKind::StringInterp { parts } = &e.kind {
+                // PROVEN iff desugarable AND every synthetic to_string is linkable. The
+                // synthetic-name list is empty for a non-desugarable interp (so the all()
+                // is vacuously true) — guard on desugarability explicitly. Only the dotted
+                // `<module>.to_string` names matter for linkability (`__str_concat` is
+                // always registered); checking the full list is sound (it contains it).
+                let desugarable = almide_mir::lower::interp_str_desugarable(parts, self.registry);
+                let all_linkable = almide_mir::lower::interp_synthetic_call_names(parts, self.registry)
+                    .iter()
+                    .all(|n| !n.contains('.') || self.linkable.contains(n));
+                if desugarable && all_linkable {
+                    self.proven += 1;
+                } else {
+                    self.walled += 1;
+                }
+            }
+            almide_ir::visit::walk_expr(self, e);
+        }
+    }
+    let mut c = InterpCounter { proven: 0, walled: 0, linkable, registry };
+    almide_ir::visit::IrVisitor::visit_expr(&mut c, body);
+    (c.proven, c.walled)
+}
+
+/// Group an `Unsupported` reason into a stable histogram key: the leading clause
+/// before the first variable-debug fragment (`:`, `(`, `{`). Keeps "no scalar
+/// Repr for Named { .. }" and "no scalar Repr for Tuple [..]" in one bucket, so
+/// the histogram tracks language FEATURES, not incidental type spellings.
+fn reason_key(reason: &str) -> String {
+    reason
+        .split(|c| c == ':' || c == '(' || c == '{')
+        .next()
+        .unwrap_or(reason)
+        .trim()
+        .to_string()
+}
+
+#[derive(Default)]
+struct Tally {
+    files: usize,
+    frontend_rejected: usize,
+    frontend_panicked: usize,
+    functions: usize,
+    in_profile: usize,
+    unsupported: BTreeMap<String, usize>,
+    /// Walled functions split by category (the honest wall=0 end state):
+    ///   - `walled_real`: a pure/WASI-able function the v1 compiler cannot YET lower — THIS is the
+    ///     number the lowering-wall=0 goal drives to 0.
+    ///   - `walled_native_ffi`: a function that transitively reaches a STRUCTURAL native root
+    ///     (`@extern(rust/rs)` or a no-wasm stdlib effect) — can NEVER lower to wasm; EXCLUDED from
+    ///     the wall=0 metric (like the `@extern(wasm)` WASI imports already are). NOT a lowering bug.
+    /// Their sum equals the total `unsupported` count.
+    walled_real: usize,
+    walled_native_ffi: usize,
+    lower_panics: Vec<String>,
+    /// Functions whose certificate has an UNBACKED `+1` (the borrow-by-default
+    /// soundness gate). Must stay empty — a non-empty list is a wall breach.
+    cert_backing_breaches: Vec<String>,
+    /// Functions whose MIR call-op count EXCEEDS their source call-node count — the
+    /// caps-soundness gate for the elided-call effect markers (`record_elided_calls`).
+    /// A marker may only ADD a call-op for a genuinely ELIDED call, so `mir_calls`
+    /// can rise at most TO `ir_calls`; `mir_calls > ir_calls` means a marker
+    /// DOUBLE-COUNTED a call already lowered — which could mask a real elision and
+    /// FALSELY de-taint a Stdout-reaching function. Must stay empty (a wall breach).
+    call_count_breaches: Vec<String>,
+    /// In-profile functions whose Stdout-freedom is provable transitively (every
+    /// `Op::CallFn` callee is Stdout-free): their empty capability witness is
+    /// emitted for the proven checker. accept ⟹ no undeclared Stdout effect.
+    caps_verified: usize,
+    /// In-profile functions that call an UNANALYZABLE callee (a walled or
+    /// cross-file user function), so their Stdout-freedom cannot be proven; their
+    /// witness is NOT emitted (honest: not falsely claimed caps-safe).
+    caps_unverified: usize,
+
+    // ── interp / call-link coverage visibility metric (a/b/c) ────────────────────
+    // Every StringInterp SITE + every dotted stdlib CallFn in the corpus, bucketed
+    // HONESTLY against what the render-side wall (`try_render_wasm_program`) guarantees:
+    /// (a) PROVEN: a lowerable interp in a lowered function — it folds to a registered
+    /// `__str_concat` / `int.to_string` chain (byte-match v0 by the render_wasm
+    /// detectors). The interp's synthetic to_string callee (`int.to_string`) is in the
+    /// registry, so it renders to VALID wasm. (Interp sites only; the link metric below
+    /// covers the broader unlinkable-stdlib-call surface.)
+    interp_lowered: usize,
+    /// (b) cleanly WALLED — acceptable (no output, but never invalid wasm). Three honest
+    /// sources, all caught BEFORE any invalid module ships:
+    ///   • a non-subset interp stays the sound Opaque fallback (its fn lowers, the interp
+    ///     just defers — no synthetic call emitted);
+    ///   • an interp inside a function the lowering returned `Unsupported` for;
+    ///   • a function that emits an UNLINKABLE dotted CallFn (a stdlib fn with no
+    ///     self-host registry definition, e.g. `string.to_upper`): the render-side
+    ///     `try_render_wasm_program` REJECTS the whole program with `LowerError::
+    ///     Unsupported` rather than emit a dangling `(call $…)`. Conservative, loud, sound.
+    interp_walled: usize,
+    /// The DISTINCT unlinkable dotted stdlib callees seen across the corpus — the visible
+    /// "would be walled at render" frontier (the remaining self-host-registry gap). Folded
+    /// into bucket (b): every using program is cleanly REJECTED by the render wall, never
+    /// silently miscompiled. This is a MEASUREMENT (the gap is visible every build), NOT a
+    /// soundness breach — the wall converts each of these from a (c) hole into a (b) reject.
+    would_wall_callees: BTreeMap<String, usize>,
+    /// (c) FORBIDDEN — a silent-miscompile / invalid-wasm-passing-as-Ok: a call that
+    /// renders to a dangling `(call $…)` AND is presented as `Ok`. The render-side wall
+    /// (`try_render_wasm_program`) catches EVERY unlinkable `Op::CallFn` before output, so
+    /// no such site can escape to an `Ok` invalid module — this is 0 BY CONSTRUCTION. The
+    /// list records any site that would escape the wall (a wall-completeness breach); it
+    /// MUST stay empty.
+    forbidden_unwalled: Vec<String>,
+}
+
+/// All self-host registry CALL-NAMES (the `module.func` a `CallFn` resolves to once the
+/// v1 linker auto-includes + renames the impl) plus the always-linked `print_str`. A
+/// `CallFn` to any of these is render-resolvable. Built once from the single-source
+/// registry so it can never drift from what render_program / lower_source actually link.
+fn auto_linkable_call_names() -> HashSet<String> {
+    let mut s: HashSet<String> = almide_mir::render_wasm::self_host_runtime()
+        .iter()
+        .flat_map(|(_, entries)| entries.iter().map(|(_, call)| call.to_string()))
+        .collect();
+    s.insert("print_str".to_string());
+    s
+}
+
+/// Recursively collect `.almd` files under a path (file or directory).
+fn collect_almd(path: &Path, out: &mut Vec<PathBuf>) {
+    if path.is_dir() {
+        let mut entries: Vec<_> = match std::fs::read_dir(path) {
+            Ok(rd) => rd.filter_map(|e| e.ok().map(|e| e.path())).collect(),
+            Err(_) => return,
+        };
+        entries.sort();
+        for e in entries {
+            collect_almd(&e, out);
+        }
+    } else if path.extension().is_some_and(|x| x == "almd") {
+        out.push(path.to_path_buf());
+    }
+}
+
+/// The read-only shared inputs [`classify_lower_one_fn`] threads through the
+/// Pass-1 per-function lowering loop of [`classify_file`] — bundled to avoid a
+/// 9-parameter helper signature (the max-params trap: see
+/// docs/roadmap/active/code-health-codopsy.md).
+struct FileCtx<'a> {
+    file: &'a std::path::Path,
+    globals: &'a std::collections::HashMap<almide_ir::VarId, almide_lang::types::Ty>,
+    global_inits: &'a std::collections::HashMap<almide_ir::VarId, almide_ir::IrExpr>,
+    record_layouts: &'a almide_mir::lower::RecordLayouts,
+    variant_layouts: &'a almide_mir::lower::VariantLayouts,
+    auto_linkable: &'a HashSet<String>,
+    file_fn_names: &'a HashSet<String>,
+    native_ffi_set: &'a HashSet<String>,
+}
